partial julia implementation
this branch moves matrix construction and solving into julia. the eigensolving package is Arpack.jl.

This is not a complete implementation yet, only some of scalar mode solving functions have been ported.

From my testing, the julia implementation solves the circular fiber example in the ipynb around 36% faster than the old implementation.

Lastly, this is not plug-and-play. The fe_solver module is being renamed FEsolver